+++ cmd/audit.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"password-manager-cli/internal/core"
+	"password-manager-cli/internal/crypto"
 	"password-manager-cli/internal/storage"
 )
 
@@ -22,6 +23,7 @@ var auditCmd = &cobra.Command{
 		if err != nil {
 			return
 		}
+		defer crypto.ZeroBytes(masterPw)
 
 		vault, err := storage.LoadVault(path, masterPw)
 		if err != nil {