+++ cmd/import.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"password-manager-cli/internal/core"
+	"password-manager-cli/internal/crypto"
 	"password-manager-cli/internal/storage"
 )
 
@@ -34,6 +35,7 @@ var importCmd = &cobra.Command{
 		if err != nil {
 			return
 		}
+		defer crypto.ZeroBytes(masterPw)
 
 		vault, err := storage.LoadVault(path, masterPw)
 		if err != nil {